Close DLL's when Vcc closes; consolidate logger
This addresses issue #47;  Cartridge modeless dialogs do not close
when Vcc closes.

Fix was to call UnloadDLL in Vcc main window when close event occurs.
Previously the unload was being attempted after the execution loop
exited which was failing because resources held by some dll's (mpi.dll)
were preventing loop exit.

Also consolidated acia logger to use main Vcc logger.  Logger is only
used for debugging Vcc code. Consolidation permits uniform logging
with modules.
