+++ acia/acia.vcxproj
@@ -159,7 +159,7 @@
       <PrecompiledHeaderOutputFile Condition="'$(Configuration)|$(Platform)'=='Release|Win32'">
       </PrecompiledHeaderOutputFile>
     </ClCompile>
-    <ClCompile Include="logger.c" />
+    <ClCompile Include="..\logger.c" />
     <ClCompile Include="console.c" />
     <ClCompile Include="sc6551.c" />
     <ClCompile Include="file.c" />
